# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Bora Jogar API
  version: 0.1.0
paths:
...
  /api/v1/auth/google:
    get:
      summary: Start Google OAuth login
      parameters:
        - in: query
          name: invitation
          schema: { type: string }
      responses: { '302': { description: Redirect to Google } }
...
  /api/v1/auth/logout:
    post:
      summary: End current session
      responses: { '204': { description: Session ended }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/v1/me:
    get:
      summary: Get current user
      responses:
        '200': { $ref: '#/components/responses/CurrentUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/me/onboarding:
    get: { summary: Get onboarding progress, security: [{ cookieAuth: [] }], responses: { '200': { $ref: '#/components/responses/OnboardingProgress' } } }
    put:
      summary: Save onboarding progress
      security: [{ cookieAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OnboardingProgress' } } } }
      responses: { '200': { $ref: '#/components/responses/OnboardingProgress' }, '422': { $ref: '#/components/responses/ValidationError' } }
...
  /api/v1/admin/invitations:
    get: { summary: List invitations, responses: { '200': { description: Invitation list } } }
    post: { summary: Create invitation, responses: { '201': { description: Created }, '403': { $ref: '#/components/responses/Forbidden' } } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: borajogar_session }
  schemas:
...
    OnboardingProgress:
      type: object
      required: [currentStep, completedSteps]
      properties:
        currentStep: { type: integer, minimum: 0, maximum: 8 }
        completedSteps: { type: array, items: { type: integer, minimum: 0, maximum: 8 } }
...
    ApiError:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
        requestId: { type: string }
  responses:
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Forbidden: { description: Administrator access required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    CurrentUser: { description: Current authenticated user }
...
    OnboardingProgress: { description: Onboarding progress }
...
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }